Cadastro de usuário funcionando

+++ src/main/java/com/br/freelatech/models/Perfil.java
@@ -11,7 +11,7 @@ public class Perfil {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long perfil_id;
+	private Integer perfil_id;
 	
 	private String biografia;
 	private String social;
@@ -19,13 +19,14 @@ public class Perfil {
 	
 	@OneToOne
 	@JsonBackReference
+	//@JoinColumn(name = "usuario_id", referencedColumnName = "usuario_id")
 	private Usuario usuario;
 
-	public Long getPerfil_id() {
+	public Integer getPerfil_id() {
 		return perfil_id;
 	}
 
-	public void setPerfil_id(Long perfil_id) {
+	public void setPerfil_id(Integer perfil_id) {
 		this.perfil_id = perfil_id;
 	}
